Codex CSM: Changes
Version 108:

Finalised the profile and stat modifiers for Khorne Bezerkers. Closes
#603

Made shared entry groups for Melee and Ranged weapons, corrected the
aspiring champions using the old system.

Begun adding Black Legion and Crimson Slaughter artifacts and
restrictions as a temporary measure to the PoP time delay.

Upon completion, the codex should allow for the army supplements to be
selected in the core catalogue (allowing for access to the IA 13 rules
and profiles as well as the newer profiles).