Chaos Cultists fix
Version 126 CSM:

Added a Cultists in total entry to control the Heavy weapons cultist
count without using fixed modifiers.

Set the max modifier for both the autogun and autopistol entries to
increase as the total count is in effect. Autopistols min selection
increases with the unit total.

Heavy weapons will continue to decrease max available to autoguns and
autopistols. Autoguns will continue to decrease max available to
autopistols, with autopistols doing the same. Heavy weapons and Autoguns
will continue to decrease the minimum autopistols required in the unit.

Fixes and closes #733