Chaos fixes
Version 138:

Fixed a cultist profile entry issue, as well a cost fix in relation to
the heavy weapons cultist. Closes #814

Modifiers for the Marks of Chaos in cultists now go off the Cultists in
unit entry in an attempt to fix #821. Open pending further testing as to
see if it fixes the issue.