Chaos Supplements: Profile Overhaul Project
Crimson Slaughter and Black Legion books have been adjusted using the
Codex: CSM as a base. This should add the Forgeworld and Dataslate
units, as well as use the most recent entry layouts for the Supplements.

Any further updates will be sparingly, as the CSM codex has included
both the rules and relics of these supplements. Any serious bugs (that
prevents the catalogue from working as the supplement intends to)will be
fixed if need be.