Codex: Khorne Daemonkin
First release of Khorne Daemonkin (Version 4 Internally)

Added all unit entries to the catalogue.
Added most Wargear entries to the Catalogue. Missing Heavy Weapons.

Added the Slaughtercult as a Formation and a Core (per the upcoming
Bloodhost detachment).

This is the first release of Khorne Daemonkin. There is still plenty of
work that will need to be done.