Daemonkin minor updates
Version 12:
Changed a few profiles around in order to streamline options in
preparation for adding the Main BSB profile entries. Gave correct axe of
Khorne and Hellblade profiles to each of the respective Daemonic units.

Added Goredrinker to the Daemon Prince. Closes #870

Closing the following options based on earlier fixes.
close #864 as addressed in version 10
close #868 as addressed in version 10