Khorne Daemonkin fixes
Version 18:
Fixed a rules issue with Chaos Lord, added profile modifier for the save
when the Sigil of Corruption is selected. Closes #888
Allowed for the specific combi and power weapons to be selected for
terminators. Closes #889
Added Modifiers to the Daemon Princes save stat to account for armour.
Closes #890
Altered the bullet points in Blood for the Blood God! to numbered points
for smartphone support. Closes #891
Fixed a cost issue in the terminator champion. Closes #893
Fixed a hide issue on the Juggernaut of Khorne. Closes #895
Moved the Aura of Dark Glory to the Chaos Lord entry. Closes #901
Fixes also closes #897