Daemonkin Fixes
Version 40:

Fixed a bug involving the reaper autocannon on the Defiler. Closes #1202

Added an uncatgorised entry called Primary Detachment. When Selected,
the force with the entry will be required to Select a Warlord (and will
show Warlord). Works for Both Combined Arms and Blood Host Detachments.
Categorising the entry as "Primary Detachment" prevents it working with
the Blood Host Detachment.

Entry fixes Warlord errors and closes #1204