Tyrant Guard fixes
Moved the conditional from Tyrant Guard link into the Shared Entry as a Hide conditional.  This hides the option until Swarmlord is taken, which satisfies the quirk with IOS and still works on windows.  Also changed the entry type to No Force Org Slot instead of "No Selection" to keep the categories at a minimum.

Due to what I *think* is where link entries are stored, this change caused a ripple effect and moved every single line up 1 or 2, causing 18,900 overall changes.  I have a backup of the old .cat file in case something dies.

Closes #2166
... again