Crozius Arcanum Change
Made it easier to replace Crozius Arcanum with Chapter Relics, and moved
the Angel of Sacrifice to replace a Crozius rather than bolt pistol.
It'll look funny for any non-Imperial Fists Chaplains, because there
will be an entry group with only one option, but it makes the Angel work
properly. This will also make it easy for any future Crozius-replacement
relics.