Add files via upload
I am sure there are a few issues, but I only know of one for sure.  In the Malleus Kill Team Formation, there is no validation of the thunder hammer/ heavy thunder hammer.  With both the Vets, and the Termies able to take them, I saw no way to have a must take 2 via modifiers.