SM v2040
Added an exteremely hot gun to the command rhino.  It takes up some transport capacity though.  Fixes issues for SM noted in #3597

Sometimes a Drop Pod isn't a normal Drop Pod.  It's actually a troop. We treat.  Crazy world we live in, where a machine can be equal to an imperium soldier.  Also fixed Techmarine. Closes #3554

Unhid the Mastodon.  I figure it's spent enough time in the bushes hiding.  Closes #3593